细节交互

+++ app/src/main/java/cn/com/bluemoon/delivery/app/api/EvidenceCashApi.java
@@ -63,7 +63,7 @@ public class EvidenceCashApi extends DeliveryApi {
      * @param token token身份检验码 String
      * @param type 操作方式 String
      */
-    public static void saveCashInfo(double cashAmount,String evidencePath,String token,String type,AsyncHttpResponseHandler handler){
+    public static void saveCashInfo(long cashAmount,String evidencePath,String token,String type,AsyncHttpResponseHandler handler){
         if(null == token||null == type) {
             return;
         }
